requirements.txt
# Consolidated project requirements (single source of truth)

## Core (Flask app)
flask>=2.3.0                 # Core web framework
sqlalchemy>=2.0.0            # Database ORM
flask-sqlalchemy>=3.0.0      # Flask integration for SQLAlchemy
flask-migrate                # Database migrations
flask-login                  # User authentication
flask-wtf                    # Form handling
python-dotenv>=1.0.0         # Environment variable loading
werkzeug                     # WSGI utilities
psycopg2-binary>=2.9.0       # PostgreSQL driver

## MCP server and async stack
rdflib>=7.0.0                # RDF graph processing for ontology support
aiohttp>=3.9.0               # Async HTTP server for MCP server
mcp[cli]                     # Model Context Protocol
httpx>=0.26.0                # HTTP client used by MCP
websockets>=12.0.0           # WebSockets for MCP server
pydantic>=2.6.0              # Data validation for MCP server
fastapi>=0.108.0             # FastAPI for MCP server
uvicorn>=0.24.0              # ASGI server for FastAPI
sparqlwrapper>=2.0.0         # SPARQL querying for ontologies

## API communication
requests>=2.31.0             # HTTP client for API requests

## Production
gunicorn                     # WSGI HTTP Server

## NLP
nltk                          # Natural language processing

## Testing & development
pytest
pytest-flask
email_validator
black
flake8

## LLM providers
anthropic==0.51.0            # Anthropic Python SDK (latest verified)
openai>=1.12.0               # OpenAI API client
google-generativeai>=0.7.2   # Google Gemini API client

## LangChain / LangGraph
langchain
langchain-core
langchain-anthropic
langchain-community
langgraph

## Document processing
PyPDF2                       # PDF text extraction
python-docx                  # DOCX text extraction
beautifulsoup4>=4.12.0       # HTML text extraction
markdown                     # Markdown to HTML conversion
aiofiles>=23.2.0             # Async file operations

## Vector embeddings
pgvector                     # PostgreSQL vector extension
sentence-transformers        # For generating embeddings
chromadb                     # Vector database

## Data science & ML
numpy>=1.26.0                # Numerical computing
pandas>=2.2.0                # Data manipulation
scikit-learn>=1.4.0          # Machine learning toolkit
tiktoken>=0.5.2              # Token counting for OpenAI models

## Async utilities
typing-extensions>=4.9.0     # Type hints
tenacity>=8.2.0              # Retry logic
nest-asyncio>=1.6.0          # Nested async event loops

## Zotero integration
pyzotero                      # Zotero API client
python-slugify==8.0.4

## Visualization (optional)
matplotlib>=3.8.0            # Plotting library
networkx>=3.2.0              # Network analysis

## Graph database
neo4j==5.28.2                # Official Neo4j Python driver

## System metrics
psutil                       # Used by admin system health diagnostics
